+++ src/pages/index/calendar/calendar.vue
@@ -1,0 +1,8 @@
+<script setup lang="ts"></script>
+
+<template>
+  <UBasePage>
+    <!-- 标题栏 -->
+    <template v-slot:navContent>校历</template>
+  </UBasePage>
+</template>